unsigned long vs. int fix

+++ src/wrench/simgrid_S4U_util/S4U_PendingCommunication.cpp
@@ -89,10 +89,10 @@ namespace wrench {
             pending_s4u_comms.push_back((*it)->comm_ptr);
         }
 
-        unsigned long index = 0;
+        int index = 0;
         bool one_comm_failed = false;
         try {
-            index = (unsigned long) simgrid::s4u::Comm::wait_any_for(&pending_s4u_comms, timeout);
+            index =  simgrid::s4u::Comm::wait_any_for(&pending_s4u_comms, timeout);
 #ifdef MESSAGE_MANAGER
             MessageManager::removeReceivedMessage(pending_comms[index]->mailbox_name, pending_comms[index]->simulation_message.get());
 #endif
@@ -118,7 +118,7 @@ namespace wrench {
             }
         }
 
-        return index;
+        return (unsigned long)index;
     }
 
 };
